Rename StjAttributesTable and expose it publicly.
Also:
- System.Text.Json reference is high enough that we can use the more efficient deserialization from JsonElement.
- Recreate the extension methods as instance methods on the class, and mark the extension methods themselves as obsolete.
- Add some parameter validation for situations that we previously assumed could never happen.
