+++ src/NetTopologySuite.IO.GeoJSON4STJ/Converters/StjAttributesTableConverter.cs
@@ -21,7 +21,7 @@ namespace NetTopologySuite.IO.Converters
         /// <summary>
         /// Writes the JSON representation of the object.
         /// </summary>
-        /// <param name="writer">The <see cref="T:Newtonsoft.Json.JsonWriter"/> to write to.</param>
+        /// <param name="writer">The <see cref="Utf8JsonWriter"/> to write to.</param>
         /// <param name="value">The value.</param>
         /// <param name="options">The calling serializer.</param>
         public override void Write(Utf8JsonWriter writer, IAttributesTable value, JsonSerializerOptions options)
@@ -49,7 +49,7 @@ namespace NetTopologySuite.IO.Converters
         /// <summary>
         /// Reads the JSON representation of the object.
         /// </summary>
-        /// <param name="reader">The <see cref="T:Newtonsoft.Json.JsonReader"/> to read from.</param>
+        /// <param name="reader">The <see cref="Utf8JsonReader"/> to read from.</param>
         /// <param name="objectType">Type of the object.</param>
         /// <param name="options">The calling serializer.</param>
         /// <returns>
@@ -66,7 +66,7 @@ namespace NetTopologySuite.IO.Converters
                         return null;
 
                     case JsonValueKind.Object:
-                        return new StjAttributesTable(doc.RootElement);
+                        return new JsonElementAttributesTable(doc.RootElement);
 
                     default:
                         throw new JsonException();
