+++ test/NetTopologySuite.IO.GeoJSON4STJ.Test/Converters/StjAttributesTableExtensionsTests.cs
@@ -9,6 +9,10 @@ using NUnit.Framework;
 
 namespace NetTopologySuite.IO.GeoJSON4STJ.Test.Converters
 {
+    // These methods were moved to instance methods on JsonElementAttributesTable, but the extension
+    // methods need to remain for compat reasons, so:
+    #pragma warning disable CS0618
+
     public sealed class StjAttributesTableExtensionsTests
     {
         [Test]
@@ -285,6 +289,10 @@ namespace NetTopologySuite.IO.GeoJSON4STJ.Test.Converters
                 Assert.That(!feature.Attributes.TryGetJsonObjectPropertyValue("null", options, out DateTime _), () => "DateTime from null");
                 Assert.That(!feature.Attributes.TryGetJsonObjectPropertyValue("null", options, out DateTimeOffset _), () => "DateTimeOffset from null");
                 Assert.That(!feature.Attributes.TryGetJsonObjectPropertyValue("null", options, out Guid _), () => "Guid from null");
+
+                // as of 3.x, we now expose the table so that the caller can get the JsonElement for
+                // their own inspection and manipulation.
+                Assert.That(feature.Attributes, Is.InstanceOf<JsonElementAttributesTable>());
             });
 
             // complex type, with two properties: one of a user-defined complex type, and one that
